# opt: optimize FuncDepSet.ReduceCols