@mgartner 2023-10-30
Slow query benchmark results:

```
name                         old time/op    new time/op    delta
SlowQueries/slow-query-1-24    28.9ms ± 1%    28.1ms ± 1%   -2.92%  (p=0.008 n=5+5)
SlowQueries/slow-query-2-24     418ms ± 1%     419ms ± 1%     ~     (p=0.841 n=5+5)
SlowQueries/slow-query-3-24    97.7ms ± 1%    97.5ms ± 1%     ~     (p=0.841 n=5+5)
SlowQueries/slow-query-4-24     697ms ± 0%     698ms ± 0%     ~     (p=0.841 n=5+5)
SlowQueries/slow-query-5-24     4.35s ± 0%     1.55s ± 1%  -64.40%  (p=0.008 n=5+5)

name                         old alloc/op   new alloc/op   delta
SlowQueries/slow-query-1-24    6.83MB ± 0%    6.81MB ± 0%   -0.25%  (p=0.008 n=5+5)
SlowQueries/slow-query-2-24    48.5MB ± 0%    48.6MB ± 0%     ~     (p=0.222 n=5+5)
SlowQueries/slow-query-3-24    44.8MB ± 0%    44.6MB ± 0%   -0.28%  (p=0.008 n=5+5)
SlowQueries/slow-query-4-24     153MB ± 0%     153MB ± 0%   -0.19%  (p=0.008 n=5+5)
SlowQueries/slow-query-5-24     405MB ± 0%     290MB ± 0%  -28.43%  (p=0.029 n=4+4)

name                         old allocs/op  new allocs/op  delta
SlowQueries/slow-query-1-24     70.2k ± 0%     69.7k ± 0%   -0.75%  (p=0.008 n=5+5)
SlowQueries/slow-query-2-24      391k ± 0%      391k ± 0%     ~     (p=0.151 n=5+5)
SlowQueries/slow-query-3-24      389k ± 0%      385k ± 0%   -1.01%  (p=0.008 n=5+5)
SlowQueries/slow-query-4-24     3.23M ± 0%     3.22M ± 0%   -0.29%  (p=0.008 n=5+5)
SlowQueries/slow-query-5-24     8.94M ± 0%     5.34M ± 0%  -40.28%  (p=0.008 n=5+5)
```

I'll run the BenchmarkPhases/Explore benchmarks and post them soon.