@mgartner 2023-10-31
I made the copy within `ReduceCols` lazy to avoid more allocations. Benchmark results below.

```
Gathered from:
./dev bench pkg/sql/opt/bench -f BenchmarkSlow --cpus 1 --count 5 --verbose --ignore-cache --bench-mem --timeout 3000s

name                         old time/op    new time/op    delta
SlowQueries/slow-query-1-24    28.1ms ± 1%    27.3ms ± 2%   -2.96%  (p=0.008 n=5+5)
SlowQueries/slow-query-2-24     402ms ± 1%     400ms ± 0%     ~     (p=0.151 n=5+5)
SlowQueries/slow-query-3-24    95.8ms ± 1%    95.1ms ± 1%     ~     (p=0.056 n=5+5)
SlowQueries/slow-query-4-24     675ms ± 0%     674ms ± 0%     ~     (p=0.421 n=5+5)
SlowQueries/slow-query-5-24     4.33s ± 1%     1.51s ± 1%  -65.09%  (p=0.008 n=5+5)

name                         old alloc/op   new alloc/op   delta
SlowQueries/slow-query-1-24    6.83MB ± 0%    6.81MB ± 0%   -0.26%  (p=0.008 n=5+5)
SlowQueries/slow-query-2-24    48.6MB ± 0%    48.6MB ± 0%     ~     (p=0.548 n=5+5)
SlowQueries/slow-query-3-24    44.8MB ± 0%    44.6MB ± 0%   -0.31%  (p=0.008 n=5+5)
SlowQueries/slow-query-4-24     153MB ± 0%     153MB ± 0%   -0.24%  (p=0.008 n=5+5)
SlowQueries/slow-query-5-24     405MB ± 0%     283MB ± 0%  -30.23%  (p=0.029 n=4+4)

name                         old allocs/op  new allocs/op  delta
SlowQueries/slow-query-1-24     70.2k ± 0%     69.6k ± 0%   -0.79%  (p=0.008 n=5+5)
SlowQueries/slow-query-2-24      391k ± 0%      391k ± 0%     ~     (p=0.421 n=5+5)
SlowQueries/slow-query-3-24      389k ± 0%      385k ± 0%   -1.11%  (p=0.008 n=5+5)
SlowQueries/slow-query-4-24     3.23M ± 0%     3.22M ± 0%   -0.35%  (p=0.008 n=5+5)
SlowQueries/slow-query-5-24     8.94M ± 0%     5.11M ± 0%  -42.83%  (p=0.008 n=5+5)
```

```
Gathered from:
./dev bench pkg/sql/opt/bench -f 'BenchmarkPhases/.*/.*/Explore' --cpus 1 --count 5 --verbose --ignore-cache --bench-mem --timeout 3000s

name                                                    old time/op    new time/op    delta
Phases/kv-read/Simple/Explore-24                          50.9µs ± 0%    50.5µs ± 1%     ~     (p=0.690 n=5+5)
Phases/kv-read-const/Simple/Explore-24                    51.3µs ± 1%    50.6µs ± 1%   -1.48%  (p=0.008 n=5+5)
Phases/tpcc-new-order/Simple/Explore-24                    135µs ± 1%     137µs ± 2%   +1.43%  (p=0.032 n=5+5)
Phases/tpcc-delivery/Simple/Explore-24                     114µs ± 1%     114µs ± 3%     ~     (p=0.421 n=5+5)
Phases/tpcc-delivery/Prepared/Explore-24                  42.8µs ± 0%    42.8µs ± 1%     ~     (p=1.000 n=5+5)
Phases/tpcc-stock-level/Simple/Explore-24                  604µs ± 1%     599µs ± 1%     ~     (p=0.095 n=5+5)
Phases/tpcc-stock-level/Prepared/Explore-24                261µs ± 0%     258µs ± 0%   -0.94%  (p=0.008 n=5+5)
Phases/many-columns-and-indexes-a/Simple/Explore-24        453µs ± 1%     471µs ± 0%   +3.97%  (p=0.008 n=5+5)
Phases/many-columns-and-indexes-a/Prepared/Explore-24      399µs ± 0%     418µs ± 0%   +4.87%  (p=0.008 n=5+5)
Phases/many-columns-and-indexes-b/Simple/Explore-24        539µs ± 0%     560µs ± 0%   +3.91%  (p=0.008 n=5+5)
Phases/many-columns-and-indexes-b/Prepared/Explore-24      452µs ± 0%     472µs ± 1%   +4.44%  (p=0.008 n=5+5)
Phases/many-columns-and-indexes-c/Simple/Explore-24        324ms ± 2%     299ms ± 1%   -7.90%  (p=0.008 n=5+5)
Phases/many-columns-and-indexes-c/Prepared/Explore-24      337ms ± 4%     298ms ± 0%  -11.51%  (p=0.008 n=5+5)
Phases/single-col-histogram-range/Simple/Explore-24        138µs ± 1%     137µs ± 0%     ~     (p=0.310 n=5+5)
Phases/single-col-histogram-range/Prepared/Explore-24     58.7µs ± 0%    58.4µs ± 0%   -0.49%  (p=0.032 n=5+5)
Phases/batch-insert-one/Simple/Explore-24                  115µs ± 1%     115µs ± 1%     ~     (p=0.310 n=5+5)
Phases/batch-insert-many/Simple/Explore-24                8.46ms ± 1%    8.46ms ± 1%     ~     (p=1.000 n=5+5)
Phases/ored-preds-100/Simple/Explore-24                   3.69ms ± 1%    3.68ms ± 0%     ~     (p=0.421 n=5+5)
Phases/ored-preds-using-params-100/Simple/Explore-24      3.48ms ± 0%    3.47ms ± 0%   -0.29%  (p=0.016 n=5+4)
Phases/ored-preds-using-params-100/Prepared/Explore-24    1.52ms ± 1%    1.56ms ± 3%     ~     (p=0.151 n=5+5)

name                                                    old alloc/op   new alloc/op   delta
Phases/kv-read/Simple/Explore-24                          15.4kB ± 0%    15.4kB ± 0%     ~     (p=1.000 n=5+5)
Phases/kv-read-const/Simple/Explore-24                    15.5kB ± 0%    15.5kB ± 0%     ~     (p=1.000 n=5+5)
Phases/tpcc-new-order/Simple/Explore-24                   32.7kB ± 0%    32.7kB ± 0%     ~     (p=0.460 n=5+5)
Phases/tpcc-delivery/Simple/Explore-24                    26.7kB ± 0%    26.7kB ± 0%     ~     (p=0.381 n=5+5)
Phases/tpcc-delivery/Prepared/Explore-24                  12.8kB ± 0%    12.8kB ± 0%     ~     (p=1.000 n=5+5)
Phases/tpcc-stock-level/Simple/Explore-24                  130kB ± 0%     130kB ± 0%   +0.01%  (p=0.032 n=5+5)
Phases/tpcc-stock-level/Prepared/Explore-24               58.9kB ± 0%    58.9kB ± 0%     ~     (p=0.611 n=5+5)
Phases/many-columns-and-indexes-a/Simple/Explore-24       39.7kB ± 0%    39.7kB ± 0%     ~     (p=0.183 n=5+5)
Phases/many-columns-and-indexes-a/Prepared/Explore-24     5.50kB ± 0%    5.50kB ± 0%     ~     (all equal)
Phases/many-columns-and-indexes-b/Simple/Explore-24       47.9kB ± 0%    48.0kB ± 0%     ~     (p=0.794 n=5+5)
Phases/many-columns-and-indexes-b/Prepared/Explore-24     8.02kB ± 0%    8.02kB ± 0%     ~     (p=0.381 n=5+5)
Phases/many-columns-and-indexes-c/Simple/Explore-24        110MB ± 0%     109MB ± 0%   -1.21%  (p=0.008 n=5+5)
Phases/many-columns-and-indexes-c/Prepared/Explore-24      109MB ± 0%     108MB ± 0%   -1.21%  (p=0.008 n=5+5)
Phases/single-col-histogram-range/Simple/Explore-24       29.8kB ± 0%    29.8kB ± 0%     ~     (p=0.643 n=5+5)
Phases/single-col-histogram-range/Prepared/Explore-24     13.0kB ± 0%    13.0kB ± 0%     ~     (all equal)
Phases/batch-insert-one/Simple/Explore-24                 64.8kB ± 0%    64.8kB ± 0%     ~     (p=0.214 n=5+5)
Phases/batch-insert-many/Simple/Explore-24                3.92MB ± 0%    3.92MB ± 0%     ~     (p=0.841 n=5+5)
Phases/ored-preds-100/Simple/Explore-24                    945kB ± 0%     945kB ± 0%     ~     (p=0.595 n=5+5)
Phases/ored-preds-using-params-100/Simple/Explore-24       926kB ± 0%     926kB ± 0%     ~     (p=0.421 n=5+5)
Phases/ored-preds-using-params-100/Prepared/Explore-24     736kB ± 0%     736kB ± 0%     ~     (p=0.746 n=5+5)

name                                                    old allocs/op  new allocs/op  delta
Phases/kv-read/Simple/Explore-24                            97.0 ± 0%      97.0 ± 0%     ~     (all equal)
Phases/kv-read-const/Simple/Explore-24                      97.0 ± 0%      97.0 ± 0%     ~     (all equal)
Phases/tpcc-new-order/Simple/Explore-24                      177 ± 0%       177 ± 0%     ~     (all equal)
Phases/tpcc-delivery/Simple/Explore-24                       167 ± 0%       167 ± 0%     ~     (all equal)
Phases/tpcc-delivery/Prepared/Explore-24                    75.0 ± 0%      75.0 ± 0%     ~     (all equal)
Phases/tpcc-stock-level/Simple/Explore-24                    711 ± 0%       711 ± 0%     ~     (all equal)
Phases/tpcc-stock-level/Prepared/Explore-24                  341 ± 0%       341 ± 0%     ~     (p=0.556 n=4+5)
Phases/many-columns-and-indexes-a/Simple/Explore-24         83.0 ± 0%      83.0 ± 0%     ~     (all equal)
Phases/many-columns-and-indexes-a/Prepared/Explore-24       31.0 ± 0%      31.0 ± 0%     ~     (all equal)
Phases/many-columns-and-indexes-b/Simple/Explore-24          138 ± 0%       138 ± 0%     ~     (all equal)
Phases/many-columns-and-indexes-b/Prepared/Explore-24       49.0 ± 0%      49.0 ± 0%     ~     (all equal)
Phases/many-columns-and-indexes-c/Simple/Explore-24         762k ± 0%      721k ± 0%   -5.48%  (p=0.008 n=5+5)
Phases/many-columns-and-indexes-c/Prepared/Explore-24       756k ± 0%      715k ± 0%   -5.44%  (p=0.016 n=5+4)
Phases/single-col-histogram-range/Simple/Explore-24          524 ± 0%       524 ± 0%     ~     (all equal)
Phases/single-col-histogram-range/Prepared/Explore-24        251 ± 0%       251 ± 0%     ~     (all equal)
Phases/batch-insert-one/Simple/Explore-24                    325 ± 0%       325 ± 0%     ~     (all equal)
Phases/batch-insert-many/Simple/Explore-24                 28.5k ± 0%     28.5k ± 0%     ~     (p=0.881 n=5+5)
Phases/ored-preds-100/Simple/Explore-24                    4.53k ± 0%     4.53k ± 0%     ~     (p=1.000 n=5+5)
Phases/ored-preds-using-params-100/Simple/Explore-24       4.53k ± 0%     4.53k ± 0%     ~     (p=1.000 n=5+5)
Phases/ored-preds-using-params-100/Prepared/Explore-24     2.88k ± 0%     2.88k ± 0%     ~     (all equal)
```