@mgartner 2023-11-02
I've removed the lazy copy of `cols` within `ReduceCols`. The gains from it were marginal and I think there's a safer way to do it (thanks to @DrewKimball) — make a `ReduceColsNoCopy` version. I'll do that in a separate PR so that I don't hold this one up.

Here's the new benchmark results:


```
name                         old time/op    new time/op    delta
SlowQueries/slow-query-1-24    28.5ms ± 0%    27.6ms ± 1%   -3.25%  (p=0.016 n=4+5)
SlowQueries/slow-query-2-24     413ms ± 1%     409ms ± 1%   -0.89%  (p=0.032 n=5+5)
SlowQueries/slow-query-3-24    97.4ms ± 1%    96.6ms ± 1%     ~     (p=0.095 n=5+5)
SlowQueries/slow-query-4-24     692ms ± 1%     682ms ± 1%   -1.44%  (p=0.008 n=5+5)
SlowQueries/slow-query-5-24     4.35s ± 0%     1.52s ± 0%  -65.08%  (p=0.016 n=5+4)

name                         old alloc/op   new alloc/op   delta
SlowQueries/slow-query-1-24    6.83MB ± 0%    6.81MB ± 0%   -0.25%  (p=0.008 n=5+5)
SlowQueries/slow-query-2-24    48.6MB ± 0%    48.6MB ± 0%     ~     (p=0.421 n=5+5)
SlowQueries/slow-query-3-24    44.8MB ± 0%    44.6MB ± 0%   -0.28%  (p=0.008 n=5+5)
SlowQueries/slow-query-4-24     153MB ± 0%     153MB ± 0%   -0.20%  (p=0.008 n=5+5)
SlowQueries/slow-query-5-24     405MB ± 0%     290MB ± 0%  -28.43%  (p=0.029 n=4+4)

name                         old allocs/op  new allocs/op  delta
SlowQueries/slow-query-1-24     70.2k ± 0%     69.7k ± 0%   -0.75%  (p=0.008 n=5+5)
SlowQueries/slow-query-2-24      391k ± 0%      391k ± 0%     ~     (p=0.095 n=5+5)
SlowQueries/slow-query-3-24      389k ± 0%      385k ± 0%   -1.01%  (p=0.008 n=5+5)
SlowQueries/slow-query-4-24     3.23M ± 0%     3.22M ± 0%   -0.29%  (p=0.008 n=5+5)
SlowQueries/slow-query-5-24     8.94M ± 0%     5.34M ± 0%  -40.28%  (p=0.008 n=5+5)
```

```
name                                                    old time/op    new time/op    delta
Phases/kv-read/Simple/Explore-24                          51.1µs ± 1%    51.1µs ± 2%    ~     (p=1.000 n=5+5)
Phases/kv-read-const/Simple/Explore-24                    51.5µs ± 1%    51.1µs ± 1%    ~     (p=0.056 n=5+5)
Phases/tpcc-new-order/Simple/Explore-24                    136µs ± 1%     136µs ± 1%    ~     (p=0.310 n=5+5)
Phases/tpcc-delivery/Simple/Explore-24                     115µs ± 1%     115µs ± 0%  -0.64%  (p=0.032 n=5+5)
Phases/tpcc-delivery/Prepared/Explore-24                  43.7µs ± 1%    43.3µs ± 0%    ~     (p=0.095 n=5+5)
Phases/tpcc-stock-level/Simple/Explore-24                  605µs ± 0%     605µs ± 0%    ~     (p=0.841 n=5+5)
Phases/tpcc-stock-level/Prepared/Explore-24                263µs ± 1%     263µs ± 0%    ~     (p=0.841 n=5+5)
Phases/many-columns-and-indexes-a/Simple/Explore-24        457µs ± 1%     477µs ± 0%  +4.54%  (p=0.008 n=5+5)
Phases/many-columns-and-indexes-a/Prepared/Explore-24      401µs ± 1%     423µs ± 0%  +5.67%  (p=0.008 n=5+5)
Phases/many-columns-and-indexes-b/Simple/Explore-24        543µs ± 0%     566µs ± 1%  +4.32%  (p=0.008 n=5+5)
Phases/many-columns-and-indexes-b/Prepared/Explore-24      454µs ± 0%     479µs ± 0%  +5.44%  (p=0.008 n=5+5)
Phases/many-columns-and-indexes-c/Simple/Explore-24        331ms ± 2%     304ms ± 2%  -8.10%  (p=0.008 n=5+5)
Phases/many-columns-and-indexes-c/Prepared/Explore-24      339ms ± 1%     309ms ± 1%  -8.83%  (p=0.008 n=5+5)
Phases/single-col-histogram-range/Simple/Explore-24        138µs ± 1%     137µs ± 0%    ~     (p=0.095 n=5+5)
Phases/single-col-histogram-range/Prepared/Explore-24     58.9µs ± 0%    58.5µs ± 0%  -0.58%  (p=0.016 n=5+5)
Phases/batch-insert-one/Simple/Explore-24                  118µs ± 0%     118µs ± 1%    ~     (p=0.421 n=5+5)
Phases/batch-insert-many/Simple/Explore-24                8.64ms ± 0%    8.65ms ± 1%    ~     (p=0.548 n=5+5)
Phases/ored-preds-100/Simple/Explore-24                   3.71ms ± 0%    3.71ms ± 0%    ~     (p=0.548 n=5+5)
Phases/ored-preds-using-params-100/Simple/Explore-24      3.51ms ± 0%    3.51ms ± 1%    ~     (p=0.421 n=5+5)
Phases/ored-preds-using-params-100/Prepared/Explore-24    1.55ms ± 1%    1.55ms ± 1%    ~     (p=0.841 n=5+5)

name                                                    old alloc/op   new alloc/op   delta
Phases/kv-read/Simple/Explore-24                          15.4kB ± 0%    15.4kB ± 0%    ~     (all equal)
Phases/kv-read-const/Simple/Explore-24                    15.5kB ± 0%    15.5kB ± 0%    ~     (all equal)
Phases/tpcc-new-order/Simple/Explore-24                   32.7kB ± 0%    32.7kB ± 0%    ~     (p=0.222 n=5+5)
Phases/tpcc-delivery/Simple/Explore-24                    26.7kB ± 0%    26.7kB ± 0%    ~     (p=0.270 n=5+5)
Phases/tpcc-delivery/Prepared/Explore-24                  12.8kB ± 0%    12.8kB ± 0%    ~     (p=0.714 n=5+5)
Phases/tpcc-stock-level/Simple/Explore-24                  130kB ± 0%     130kB ± 0%    ~     (p=0.579 n=5+5)
Phases/tpcc-stock-level/Prepared/Explore-24               58.9kB ± 0%    58.9kB ± 0%    ~     (p=0.381 n=5+5)
Phases/many-columns-and-indexes-a/Simple/Explore-24       39.7kB ± 0%    39.7kB ± 0%    ~     (p=0.865 n=5+5)
Phases/many-columns-and-indexes-a/Prepared/Explore-24     5.50kB ± 0%    5.50kB ± 0%    ~     (all equal)
Phases/many-columns-and-indexes-b/Simple/Explore-24       47.9kB ± 0%    48.0kB ± 0%  +0.01%  (p=0.024 n=5+5)
Phases/many-columns-and-indexes-b/Prepared/Explore-24     8.02kB ± 0%    8.02kB ± 0%    ~     (p=0.444 n=5+5)
Phases/many-columns-and-indexes-c/Simple/Explore-24        110MB ± 0%     109MB ± 0%  -0.97%  (p=0.008 n=5+5)
Phases/many-columns-and-indexes-c/Prepared/Explore-24      109MB ± 0%     108MB ± 0%  -0.96%  (p=0.008 n=5+5)
Phases/single-col-histogram-range/Simple/Explore-24       29.8kB ± 0%    29.8kB ± 0%  +0.01%  (p=0.016 n=5+4)
Phases/single-col-histogram-range/Prepared/Explore-24     13.0kB ± 0%    13.0kB ± 0%    ~     (all equal)
Phases/batch-insert-one/Simple/Explore-24                 64.8kB ± 0%    64.8kB ± 0%    ~     (p=0.913 n=5+5)
Phases/batch-insert-many/Simple/Explore-24                3.92MB ± 0%    3.92MB ± 0%    ~     (p=0.548 n=5+5)
Phases/ored-preds-100/Simple/Explore-24                    945kB ± 0%     945kB ± 0%    ~     (p=0.310 n=5+5)
Phases/ored-preds-using-params-100/Simple/Explore-24       926kB ± 0%     926kB ± 0%    ~     (p=0.421 n=5+5)
Phases/ored-preds-using-params-100/Prepared/Explore-24     736kB ± 0%     736kB ± 0%    ~     (p=0.310 n=5+5)

name                                                    old allocs/op  new allocs/op  delta
Phases/kv-read/Simple/Explore-24                            97.0 ± 0%      97.0 ± 0%    ~     (all equal)
Phases/kv-read-const/Simple/Explore-24                      97.0 ± 0%      97.0 ± 0%    ~     (all equal)
Phases/tpcc-new-order/Simple/Explore-24                      177 ± 0%       177 ± 0%    ~     (all equal)
Phases/tpcc-delivery/Simple/Explore-24                       167 ± 0%       167 ± 0%    ~     (all equal)
Phases/tpcc-delivery/Prepared/Explore-24                    75.0 ± 0%      75.0 ± 0%    ~     (all equal)
Phases/tpcc-stock-level/Simple/Explore-24                    711 ± 0%       711 ± 0%    ~     (all equal)
Phases/tpcc-stock-level/Prepared/Explore-24                  341 ± 0%       341 ± 0%    ~     (all equal)
Phases/many-columns-and-indexes-a/Simple/Explore-24         83.0 ± 0%      83.0 ± 0%    ~     (all equal)
Phases/many-columns-and-indexes-a/Prepared/Explore-24       31.0 ± 0%      31.0 ± 0%    ~     (all equal)
Phases/many-columns-and-indexes-b/Simple/Explore-24          138 ± 0%       138 ± 0%    ~     (all equal)
Phases/many-columns-and-indexes-b/Prepared/Explore-24       49.0 ± 0%      49.0 ± 0%    ~     (all equal)
Phases/many-columns-and-indexes-c/Simple/Explore-24         762k ± 0%      729k ± 0%  -4.38%  (p=0.008 n=5+5)
Phases/many-columns-and-indexes-c/Prepared/Explore-24       756k ± 0%      724k ± 0%  -4.33%  (p=0.008 n=5+5)
Phases/single-col-histogram-range/Simple/Explore-24          524 ± 0%       524 ± 0%    ~     (all equal)
Phases/single-col-histogram-range/Prepared/Explore-24        251 ± 0%       251 ± 0%    ~     (all equal)
Phases/batch-insert-one/Simple/Explore-24                    325 ± 0%       325 ± 0%    ~     (all equal)
Phases/batch-insert-many/Simple/Explore-24                 28.5k ± 0%     28.5k ± 0%    ~     (p=0.683 n=5+5)
Phases/ored-preds-100/Simple/Explore-24                    4.53k ± 0%     4.53k ± 0%    ~     (p=0.238 n=4+5)
Phases/ored-preds-using-params-100/Simple/Explore-24       4.53k ± 0%     4.53k ± 0%    ~     (p=0.556 n=4+5)
Phases/ored-preds-using-params-100/Prepared/Explore-24     2.88k ± 0%     2.88k ± 0%    ~     (all equal)
```